requirements.txt
langchain
langchain-community
sentence-transformers
langchain-experimental
neo4j
pypdf
python-dotenv
langchain_openai
wikipedia
langchain_chroma
langchain-openai
graphrag
pymupdf

